chore(editor-state): move files into feature folders
