+++ packages/editor/packages/editor-state/src/index.ts
@@ -13,12 +13,12 @@ import compiler from './features/program-compiler/effect';
 import configEffect from './features/config-compiler/effect';
 import contextMenu from './features/menu/effect';
 import graphicHelper from './features/code-blocks/features/graphicHelper/effect';
-import editorSettings from './effects/editorSettings';
+import editorSettings from './features/editor-settings/effect';
 import projectImport from './features/project-import/effect';
 import pianoKeyboard from './features/code-blocks/features/pianoKeyboard/interaction';
 import projectExport from './features/project-export/effect';
 import viewport from './features/viewport/effect';
-import binaryAsset from './effects/binaryAssets';
+import binaryAsset from './features/binary-assets/effect';
 import runtime from './features/runtime/effect';
 import blockTypeUpdater from './features/code-blocks/features/blockTypeUpdater/effect';
 import shaderEffectsDeriver from './features/shader-effects/effect';
